backend/main/views/token_wrapping.py
from functools import wraps
from flask import request, jsonify
import jwt
from backend.main.models.user_table import User
from backend.main import app


def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None

        if 'x-access-token' in request.headers:
            token = request.headers['x-access-token']
        if not token:
            return jsonify({'message': 'Token is missing!'}), 401

        try:

            data = jwt.decode(token, app.config['SECRET_KEY'])

            current_user = User.query.filter_by(name=data['user']).first()
            print(current_user)
        except:
            return jsonify({'message': 'Token is invalid!'}), 401

        return f(current_user, *args, **kwargs)

    return decorated