requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.1.4
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.1
Flask-Login==0.4.1
flask-marshmallow==0.10.0
Flask-SQLAlchemy==2.3.2
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
jeepney==0.4
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
marshmallow==2.19.0
pycparser==2.19
pycrypto==2.6.1
pyflakes==1.6.0
PyJWT==1.7.1
pyxdg==0.25
redis==3.2.1
SecretStorage==3.1.1
six==1.12.0
SQLAlchemy==1.3.1
Werkzeug==0.14.1
